# hocode-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server Api Hocode.",
        "title": "Hocode API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "hocodevn.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/courses": {
            "get": {
                "description": "get courses \u003ca href=\"/api/v1/courses?page=1\u0026limit=5\"\u003e/api/v1/courses?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "List Courses",
                "operationId": "coursesx_list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Course"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Create Course",
                "operationId": "coursesx_create",
                "parameters": [
                    {
                        "description": "Create Course",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                }
            }
        },
        "/courses/:id": {
            "get": {
                "description": "get courses \u003ca href=\"/api/v1/courses?page=1\u0026limit=5\"\u003e/api/v1/courses?page=1\u0026limit=5\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "One Courses",
                "operationId": "coursesx_one",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "UpdateCourses Course",
                "operationId": "coursesx_update",
                "parameters": [
                    {
                        "description": "Update Course",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Delete Course",
                "operationId": "coursesx_delete",
                "parameters": [
                    {
                        "description": "Delete Course",
                        "name": "course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                }
            }
        },
        "/courses/{id}": {
            "get": {
                "description": "get courses by ID \u003ca href=\"/courses/5d86e07bfe6e2b157bd3b259\"\u003e/courses/5d86e07bfe6e2b157bd3b259\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Get Course By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Course"
                        }
                    }
                }
            }
        },
        "/courses/{id}/tasks": {
            "get": {
                "description": "Get Task By Courses ID \u003ca href=\"/api/v1/courses/5d86e07bfe6e2b157bd3b259/tasks\"\u003e/api/v1/courses/5d86e07bfe6e2b157bd3b259/tasks\u003c/a\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Get Task By Courses ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Task"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Course": {
            "type": "object",
            "properties": {
                "agree": {
                    "type": "boolean"
                },
                "background_image": {
                    "type": "string"
                },
                "course_desc": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "course_type": {
                    "type": "string"
                },
                "del": {
                    "type": "boolean"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "numbers_like": {
                    "type": "integer"
                },
                "rating": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "rating_value": {
                    "type": "number"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "description": "set status to public course.",
                    "type": "string"
                },
                "task_solved": {
                    "description": "Name            string     `json:\"name\" bson:\"name\"`",
                    "type": "object",
                    "$ref": "#/definitions/model.TaskSolved"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Task"
                    }
                },
                "timestamp": {
                    "type": "string"
                },
                "total_minitask": {
                    "type": "integer"
                },
                "user_create": {
                    "type": "string"
                }
            }
        },
...
        "model.Input": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.InputList": {
            "type": "object",
            "properties": {
                "input_name": {
                    "description": "ID             bson.ObjectId `json:\"id\" bson:\"_id,omitempty\"`",
                    "type": "string"
                },
                "input_type": {
                    "type": "string"
                }
            }
        },
        "model.MiniTask": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "code_point": {
                    "type": "integer"
                },
                "del": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "input_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.InputList"
                    }
                },
                "level": {
                    "type": "string"
                },
                "mini_task_desc": {
                    "type": "string"
                },
                "mini_task_name": {
                    "type": "string"
                },
                "name_func": {
                    "type": "string"
                },
                "numbers_doing": {
                    "type": "integer"
                },
                "output_type_func": {
                    "type": "string"
                },
                "point_unlock": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "template_code": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "unit_tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UnitTest"
                    }
                },
                "vitri": {
                    "type": "boolean"
                }
            }
        },
...
        "model.Task": {
            "type": "object",
            "properties": {
                "background_image": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "del": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "minitasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MiniTask"
                    }
                },
                "task_name": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.TaskSolved": {
            "type": "object",
            "properties": {
                "task_count": {
                    "description": "ID            bson.ObjectId `json:\"id\" bson:\"_id,omitempty\"`",
                    "type": "integer"
                },
                "tasks_complete": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.UnitTest": {
            "type": "object",
            "properties": {
                "expected_output": {
                    "type": "string"
                },
                "inputs": {
                    "description": "ID             bson.ObjectId `json:\"id\" bson:\"_id,omitempty\"`",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Input"
                    }
                }
            }
        },
...
    }
}